APKBUILD
Outside of making sure I conform this to Alpine Linux's standard for commits (mostly formatting), this is perfect.

It's some of the dirtiest fucking code and I hate it, but yay, it can be other dev's problems.

I'll do a simple test tonight and see if there is anything we are missing