APKBUILD
Libraries are linked!!!!!!! It took like 200 tries to pull this off, mainly because this is about 500 times more difficult than any APKBUILD file and what the code usually looks like.

JUCE has some of the strictest standards, so getting this close to be good enough for open source use means that I have to overcome so many obstacles.

Basically, you force Alpine's busybox to be respectful of the rules JUCE has. If you cannot play nice with what the framework wants, it will not let you proceed. This is why my code is so much more lengthy than anyone elses, its because there is so much high level code needed that makes attempting to do this near impossible