APKBUILD
Instead of inline writing things, it is much more appreciated to do this via shell scripting. Why? Because for code injection reasons, and it looks fucking ugly.

So instead of having an ugly APKBUILD file, we substantiate things via heredocs in a shell script. I don't know if I'm following shell script rules with regards to how Alpine handles things, but I believe I'm correct given it works in an Alpine container????????